- Moving the album modal component
